requirements.txt
fastapi 
uvicorn 
httpx 
python-dotenv 
openai 
pinecone
jinja2